// src/types/cards.ts
export type Card =
  | 'ac'
  | '2c'
  | '3c'
  | '4c'
  | '5c'
  | '6c'
  | '7c'
  | '8c'
  | '9c'
  | 'tc'
  | 'jc'
  | 'qc'
  | 'kc'
  | 'ad'
  | '2d'
  | '3d'
  | '4d'
  | '5d'
  | '6d'
  | '7d'
  | '8d'
  | '9d'
  | 'td'
  | 'jd'
  | 'qd'
  | 'kd'
  | 'ah'
  | '2h'
  | '3h'
  | '4h'
  | '5h'
  | '6h'
  | '7h'
  | '8h'
  | '9h'
  | 'th'
  | 'jh'
  | 'qh'
  | 'kh'
  | 'as'
  | '2s'
  | '3s'
  | '4s'
  | '5s'
  | '6s'
  | '7s'
  | '8s'
  | '9s'
  | 'ts'
  | 'js'
  | 'qs'
  | 'ks'

export type GameStatus = 'win' | 'lose' | 'draw' | 'burst' | undefined